fix(ci): split governance models/schemas + drop duplicate fumadocs ADR
PR 01 pushed `backend/app/models.py` (525 lines) and
`backend/app/schemas.py` (518) past the project's 500-line file
budget that `node scripts/check-file-lines.mjs` enforces in the
Frontend Check workflow. Split the four new governance ORM classes
into `backend/app/governance_models.py` and the seven matching
schemas into `backend/app/governance_schemas.py`. Both are
re-exported from the original modules so every existing
`from app.models import AuditEvent` / `from app.schemas import
AuditEventRead` site keeps working unchanged.

Also drops the leftover `.md` duplicate of the Fumadocs ADR; the
canonical `.mdx` (with Fumadocs frontmatter) was added in #201
alongside the old `.md`. The two collide on the same slug and
break `frontend test/docs-source.test.ts` (Vitest):

  Error: Duplicated slugs:
  decisions/2026-05-14-add-fumadocs-documentation-site

Pre-existing on `development` but blocks PR 01's CI gate, so fix
inline per `.claude/rules/general/no-pre-existing-excuse.md`.

Verification:
- `cd backend && uv run pytest -q` -> 435 passed, 1 skipped.
- `cd frontend && bun run check` -> exit 0 (file-lines OK,
  check-nesting OK).
- `cd frontend && bun run test` -> 363 passed (was 361 + 1 failed
  test-file before the fix).
